+++ BMCVFirmware/cmake/core_sources.cmake
@@ -23,6 +23,7 @@ set(BMCV_CORE_SOURCES
     quantizer.c
     scene.c
     stepped_random.c
+    sysex.c
     ui_channel.c
     ui_feedback.c
     ui_input.c
@@ -40,6 +41,7 @@ set(BMCV_DRIVER_SOURCES
     bmcv.c
     dac_adc.c
     fram.c
+    fw_update.c
     mcp.c
     midi.c
     presets.c